feature:  can view selected pages
After running a search, select the row item in the data view you wish to examine
then click on the row.  This enables the page aview

Click on the page view button to view a page.  Once page is shown you can click up or down to view the pages in the found page list